requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
openai==1.30.4
httpx==0.24.1
asyncpg==0.29.0